requirements.txt
telethon
request 
asyncio
os
re
random